requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.3.1
Django==3.1.4
django-redis==4.12.1
django-rename-app==0.1.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
install==1.3.4
kombu==5.0.2
prompt-toolkit==3.0.8
psycopg2==2.8.6
PyJWT==1.7.1
python-dotenv==0.15.0
python-http-client==3.3.1
pytz==2020.4
redis==3.5.3
sendgrid==6.4.8
sendgrid-python==0.1.1
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
vine==5.0.0
wcwidth==0.2.5
